refactor : money 타입을 Integer에서 Long으로 변경

+++ src/main/java/lotto/dto/InputMoneyDTO.java
@@ -1,7 +1,7 @@
 package lotto.dto;
 
-public record InputMoneyDTO(Integer money) {
-    public static InputMoneyDTO from(Integer money){
+public record InputMoneyDTO(Long money) {
+    public static InputMoneyDTO from(Long money){
         return new InputMoneyDTO(money);
     }
 }
